+++ backend/src/main/java/fr/sylvainjanet/tracker/journal/application/port/in/dtos/command/LogWeightMeasurementCommand.java
@@ -1,5 +1,6 @@
 package fr.sylvainjanet.tracker.journal.application.port.in.dtos.command;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record LogWeightMeasurementCommand(LocalDate date, Float weightInKg) {}
+public record LogWeightMeasurementCommand(LocalDate date, BigDecimal weightInKg) {}
