+++ backend/src/main/java/fr/sylvainjanet/tracker/journal/application/port/in/dtos/result/LogWeightMeasurementResult.java
@@ -1,5 +1,6 @@
 package fr.sylvainjanet.tracker.journal.application.port.in.dtos.result;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record LogWeightMeasurementResult(LocalDate date, Float weightInKg) {}
+public record LogWeightMeasurementResult(LocalDate date, BigDecimal weightInKg) {}
